+++ Plugins/Experimental/CMakeLists.txt
@@ -1,3 +1,5 @@
 add_plugin(Experimental
         "Experimental.cpp"
-        "Experimentals/SuppressPlayerLoginInfo.cpp")
+        "Experimentals/SuppressPlayerLoginInfo.cpp"
+        "Experimentals/AsyncLogFlush.cpp"
+)
