+++ Plugins/Experimental/Experimental.hpp
@@ -5,6 +5,7 @@
 namespace Experimental {
 
 class SuppressPlayerLoginInfo;
+class AsyncLogFlush;
 
 class Experimental : public NWNXLib::Plugin
 {
@@ -14,6 +15,7 @@ public:
 
 private:
     std::unique_ptr<SuppressPlayerLoginInfo> m_SuppressPlayerLoginInfo;
+    std::unique_ptr<AsyncLogFlush> m_AsyncLogFlush;
 };
 
 }
