+++ Plugins/Tweaks/CMakeLists.txt
@@ -14,5 +14,4 @@ add_plugin(Tweaks
     "Tweaks/DeadCreatureFiresOnAreaExit.cpp"
     "Tweaks/PreserveActionsOnDMPossess.cpp"
     "Tweaks/FixGreaterSanctuaryBug.cpp"
-    "Tweaks/AsyncLogFlush.cpp"
 )
